Fixed prependTo and appendTo behavior
This is something we have to keep in mind, because we now run inside of
an iframe. I changed the context of the $ function to the parent
document, however, when using selectors like "$el.appendTo('body'); "
jQuery doesn't use the $ function to follow the 'body' selector, making
it ignore our own context.

The correct way to use .appendTo and .prependTo methods is like this: "
$el.appendTo( $('body') ); "