Add Facebook Social Connect
Using omniauth-facebook and some implementation code of our own.
Did some research on using Singly. This service provides a nice
abstracted layer for connecting with several Social Services, but (at
the moment) does not allow us to publish to users accounts. Besides
that, we're not able to get all the user data we want from Singly.

Not implemented is the "weatherize callback" from Facebook when a user
weatherizes the Factlink application from Facebook.com.