Add firefox_extension feature toggle
Had to duplicate some code here for showing the content to a Firefox user which is not allowed to see the Firefox extension. It's not possible to set firefox as an unsupported browser depending on the feature toggle, since the browser detection is executed in (precompiled) javascript where we can not use the current_user to check the ability.

Duplicating this code was the fastest solution for now. Will expect this to be removed in ~ two weeks when we will fully support the FF extension and make it available to all FF users.

Signed-off-by: tomdev <tom@north26.nl>