Boyscouting: Cleaning up the rails console output.
Added lograge gem, which formats the output of requests in a one liner.
Replaced resque with our own fork of there current version, which fixes
a ' don't use :public use :public_folder' warning.
Added syncing of stdout while doing development, so we don't have to
wait for our errors.
Don't write our pid and ppid to a file so monit can check it, if we're
doing development.