Removed erroneous tests
They use methods deprecated by Rspec. And deprecated for good reason.

The behaviour is that you wait for the first exception, and then check
whether it isn't the exception it isn't allowed to be. This is no sane
behaviour whatsoever. Insert a divide by zero for the tested code, 
and suddenly it succeeds, for instance.

When converting to the suggested new way, it also showed how broken this
is, since we got random other exceptions. Seemed better to delete :)