Fixed small bug in konacha initializer
This thing had me wondering for quite a while @HackerOne, but it seems like the Runner#run method now takes an argument, which made Guard fail in a stupid way.

https://github.com/jfirebaugh/konacha/blob/master/lib/konacha/runner.rb#L17

In this pull request I make sure that all arguments passed to run are being passed to super as well. To make sure you guys don't have to run through the same troubles as we did.

Cheers!